authentication and task list views are working
The authentication is working but still trying to get the jwt to store in
localstorage.  I have the rendering of each list category AND all the
tasks within each list. Now I need to work on the back end routes and
database to better handle the querying/updating for tasks, and add route
middleware to block off most sections of app if not authenticated.